:sparkles: EditStudy 컴포넌트에 initialData props 적용
