+++ src/components/organism/index.ts
@@ -1,0 +1,1 @@
+export * from './EditStudy/EditStudy';
