+++ src/types/post.types.ts
@@ -1,3 +1,4 @@
+import { CommentType } from "./comment.types";
 export interface IPostType {
   status: 'Y' | 'N';
   title: string;
@@ -12,6 +13,28 @@ export interface IPostType {
   imageUrl: any;
   author: string;
   categoryName: string;
+}
+
+export interface IPostDetailType{
+  author: string;
+  authorPosition: string;
+  categoryName: string;
+  commentDTOList: CommentType[];
+  commentNum: number;
+  contactLink: string;
+  contactMethod: string;
+  content: string;
+  createdAt: string;
+  imageUrl: string;
+  likeNum: number;
+  onlineOrOffline: string;
+  period: string;
+  positionList: string[];
+  postId: number;
+  recruitNum: string;
+  status: string;
+  title: string;
+  view: number;
 }
